+++ server/src/common/filters/http-exception.filter.ts
@@ -1,0 +1,26 @@
+// src/common/filters/all-exceptions.filter.ts
+
+import {
+  ExceptionFilter,
+  Catch,
+  ArgumentsHost,
+  HttpException,
+  HttpStatus,
+} from '@nestjs/common';
+import { Response } from 'express';
+import { ApiResponse } from '../response/common.response';
+import { WinstonLoggerService } from '../logger/logger.service';
+
+@Catch(HttpException)
+export class HttpExceptionsFilter implements ExceptionFilter {
+  catch(exception: HttpException, host: ArgumentsHost) {
+    const response = host.switchToHttp().getResponse<Response>();
+
+    const statusCode = exception.getStatus();
+    const res = exception.getResponse();
+    const message = typeof res === 'string' ? res : res['message'];
+
+    const apiResponse = ApiResponse.responseWithNoContent(message);
+    response.status(statusCode).json(apiResponse);
+  }
+}
